fetch_agents_buy/agents/sheets_agent/requirements.txt
uagents>=0.22.0
uagents-core>=0.4.0
composio-langchain>=0.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
